add multer, crud product and category

+++ controllers/products.js
@@ -1,7 +1,9 @@
-const { products, productImages, categories } = require("../models")
-const { op, Op } = require("sequelize");
+const { products, productImages, categories, sequelize } = require("../models")
+const { Op } = require("sequelize");
 const fs = require("fs");
 const { templateResponse } = require("../helper/utils")
+const db = require("../models")
+
 
 module.exports = {
     getProducts: async (req, res, next) => {
@@ -27,39 +29,48 @@ module.exports = {
         }
     },
     addProducts: async (req, res, next) => {
+        const t = await db.sequelize.transaction();
         try {
+            if (req.accountData.role !== "ADMIN" && req.accountData.role !== "SPV") {
+                throw { rc: 401, message: "User is not authorized." }
+            }
             const checkName = await products.findOne({ where: { name: req.body.name } })
             if (checkName) {
-                req.files.forEach(async (value) => {
-                    fs.unlinkSync(`./${value.path}`)
-                })
+                req.files.forEach((value) => { fs.unlinkSync(`./${value.path}`) })
                 throw new Error("Product name has already been used")
             }
             const result = await products.create({
                 name: req.body.name,
                 price: parseInt(req.body.price),
                 stock: parseInt(req.body.stock),
-                categoryId: req.body.categoryId || null
-            })
+                categoryId: req.body.categoryId
+            }, { transaction: t })
 
-            req.files.forEach(async (value) => {
-                await productImages.create({
-                    productId: result.id,
-                    img: value.filename
-                })
-            })
+            const images = await req.files.map((value) => { return { productId: result.id, img: value.filename } })
+            await productImages.bulkCreate(images, { transaction: t })
+
+            await t.commit(); // Jika tidak ada error, maka akan tereksekusi
             return res.status(200).send({
                 success: true,
                 message: "Successfully added products"
             })
         } catch (error) {
             console.log(error);
+            req.files.forEach((value) => { fs.unlinkSync(`./${value.path}`) })
+            await t.rollback(); // Jika ada error, maka rollback yang tadi
             next(templateResponse(500, "Error adding product.", error.message))
         }
     },
     updateProduct: async (req, res, next) => {
+        // const t = await db.sequelize.transaction()
         try {
-            const checkProduct = await products.findOne({ where: { id: req.params.id }, include: [{ model: productImages, attributes: ["img"], raw: true }] })
+            if (req.accountData.role !== "ADMIN" && req.accountData.role !== "SPV") {
+                throw { rc: 401, message: "User is not authorized." }
+            }
+            const checkProduct = await products.findOne({
+                where: { id: req.params.id },
+                include: [{ model: productImages, attributes: ["img"], raw: true }]
+            })
             if (!checkProduct) { throw new Error("Can't find the selected product.") }
             if (req.files.length) {
                 checkProduct.productImages.forEach((value) => {
@@ -90,19 +101,23 @@ module.exports = {
 
         } catch (error) {
             console.log(error);
-            next(templateResponse(500, "Error updating product.", error.message))
+            req.files.forEach((value) => { fs.unlinkSync(`./${value.path}`) })
+            next(templateResponse(error.rc || 500, "Error updating product.", error.message))
         }
     },
     deleteProduct: async (req, res, next) => {
         try {
+            if (req.accountData.role !== "ADMIN" && req.accountData.role !== "SPV") {
+                throw { rc: 401, message: "User is not authorized." }
+            }
             const result = await products.destroy({ where: { id: req.params.id } })
             if (!result) { throw { rc: 400, message: "Cannot find selected product." } }
             else {
                 const result2 = await productImages.findAll({ where: { productId: req.params.id }, raw: true })
+                await productImages.destroy({ where: { productId: req.params.id } })
                 result2.forEach((value) => {
                     fs.unlinkSync(`./public/productImages/${value.img}`)
                 })
-                await productImages.destroy({ where: { productId: req.params.id } })
                 return res.status(200).send({ success: true, message: "Successfully deleted product." })
             }
         } catch (error) {
